Bulk print on approved expenses not opening in new tab (#645)
* add bulk printing of financial audit reports when approving expenses

* remove bulk-generator

* add safeguard in case no financial numbers get returned on approving expenses
