+++ server/routes/reporting/index.js
@@ -7,13 +7,6 @@ const { financialAudit } = require('./audit.controller');
   const { isCourtUser } = require('../../components/auth/user-type');
   const { getReports, getStatistics } = require('./reporting.controller');
 
-  // TODO: these are in for POC and demoing purposes and as a reference for the future
-  const {
-    getSomeReport,
-    generateSomeReport,
-    generateSomeOtherReport,
-  } = require('./some-report/some-report.controller');
-
   const { generateBulk } = require('./expenses-bulk/expenses-bulk.controller');
 
   module.exports = function(app) {
@@ -37,27 +30,9 @@ const { financialAudit } = require('./audit.controller');
       isCourtUser,
       getStatistics(app));
 
-    app.get('/reporting/some-report',
-      'some-report.get',
-      auth.verify,
-      getSomeReport(app));
-    app.get('/reporting/some-report/some-report.pdf',
-      'some-report.generate.get',
-      auth.verify,
-      generateSomeReport(app));
-    app.get('/reporting/some-report/some-other-report.pdf',
-      'some-other-report.generate.get',
-      auth.verify,
-      generateSomeOtherReport(app));
-
-    app.get('/reporting/expenses-bulk/expenses-bulk.pdf',
-      'expenses-bulk.generate.get',
-      generateBulk(app));
-
-    app.get('/reports/financial-audit/:auditNumber',
-      'reports.financial-audit.get',
-      auth.verify,
-      financialAudit(app));
+    // match bulk first, if it does not match go to single audit
+    app.get('/reports/financial-audit/bulk', 'reports.financial-audit.bulk.get', auth.verify, generateBulk(app));
+    app.get('/reports/financial-audit/:auditNumber', 'reports.financial-audit.get', auth.verify, financialAudit(app));
   };
 
 })();
